Add files via upload
I'm not sure exactly how to use the states so I commented them out for now. I could replace each state with a bool if that works with the UI. 
The init() function is just like a default constructor so I made up some values but feel free to change those. 
I made an Update() function that returns a bool like the pokemon class. My idea for update is this: input a time, and if amount slept is less than the time than the pet is sleepy. If the amount eaten is less than the time, the pokemon is hungry. If the pokemon is tired and hungry at the same time then it is unhealthy. If it is neither hungry nor sleepy then it is healthy.
At the end of the file I made a checkpoint to test the functions.